first version

+++ src/subsample.rs
@@ -3,12 +3,22 @@ use crate::sub_cli::SampleArgs;
 use rand::rngs::StdRng;
 use rand::{Rng, SeedableRng};
 
+/// Parses FASTQ/A file and sampling according to the seed and fraction.
+/// Outputs the results to [`std::io::stdout()`].
+///
+/// # Arguments
+///
+/// Check the arguments by `--help`
+///
+/// # Errors
+///
+/// Return an error if the operation cannot be completed.
 pub fn subsample_fastx(
     fx_path: &str,
     sparas: &SampleArgs,
     is_fasta: bool,
 ) -> Result<(), std::io::Error> {
-    let rand_seed = sparas.random_seed.unwrap_or(11);
+    let rand_seed = sparas.random_seed.unwrap_or(11) as u64;
     let sampling_frac = sparas.sample_fraction.unwrap_or(1.0);
     let mut rng = StdRng::seed_from_u64(rand_seed);
